+++ RiotSharp/MatchEndpoint/Enums/CapturedPoint.cs
@@ -1,6 +1,7 @@
 ﻿using Newtonsoft.Json;
+using RiotSharp.MatchEndpoint.Enums.Converters;
 
-namespace RiotSharp.MatchEndpoint
+namespace RiotSharp.MatchEndpoint.Enums
 {
     /// <summary>
     /// Class representing a capture point in a dominion game (Match API).
