+++ RiotSharp/MatchEndpoint/Enums/Converters/RoleConverter.cs
@@ -3,9 +3,9 @@ using Newtonsoft.Json.Linq;
 using System;
 using System.Reflection;
 
-namespace RiotSharp.MatchEndpoint
+namespace RiotSharp.MatchEndpoint.Enums.Converters
 {
-    class SeasonConverter : JsonConverter
+    class RoleConverter : JsonConverter
     {
         public override bool CanConvert(Type objectType)
         {
@@ -20,22 +20,16 @@ namespace RiotSharp.MatchEndpoint
             var str = token.Value<string>();
             switch (str)
             {
-                case "PRESEASON3":
-                    return Season.PreSeason3;
-                case "SEASON3":
-                    return Season.Season3;
-                case "PRESEASON2014":
-                    return Season.PreSeason2014;
-                case "SEASON2014":
-                    return Season.Season2014;
-                case "PRESEASON2015":
-                    return Season.PreSeason2015;
-                case "SEASON2015":
-                    return Season.Season2015;
-                case "PRESEASON2016":
-                    return Season.PreSeason2016;
-                case "SEASON2016":
-                    return Season.Season2016;
+                case "DUO":
+                    return Role.Duo;
+                case "NONE":
+                    return Role.None;
+                case "SOLO":
+                    return Role.Solo;
+                case "DUO_CARRY":
+                    return Role.DuoCarry;
+                case "DUO_SUPPORT":
+                    return Role.DuoSupport;
                 default:
                     return null;
             }
@@ -43,7 +37,7 @@ namespace RiotSharp.MatchEndpoint
 
         public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
         {
-            serializer.Serialize(writer, ((Season)value).ToString().ToUpper());
+            serializer.Serialize(writer, ((Role)value).ToCustomString());
         }
     }
 }
