+++ RiotSharp/MatchEndpoint/Enums/QueueType.cs
@@ -1,6 +1,7 @@
 ﻿using Newtonsoft.Json;
+using RiotSharp.MatchEndpoint.Enums.Converters;
 
-namespace RiotSharp.MatchEndpoint
+namespace RiotSharp.MatchEndpoint.Enums
 {
     /// <summary>
     /// Match queue type (Match API).
