+++ RiotSharp/RiotSharp.csproj
@@ -82,15 +82,15 @@
     <Compile Include="IStatusRiotApi.cs" />
     <Compile Include="ITournamentRiotApi.cs" />
     <Compile Include="MatchEndpoint\MatchList.cs" />
-    <Compile Include="MatchEndpoint\Converters\AscendedType.cs" />
-    <Compile Include="MatchEndpoint\Converters\AscendedTypeConverter.cs" />
-    <Compile Include="MatchEndpoint\Converters\CapturedPointConverter.cs" />
+    <Compile Include="MatchEndpoint\Enums\AscendedType.cs" />
+    <Compile Include="MatchEndpoint\Enums\Converters\AscendedTypeConverter.cs" />
+    <Compile Include="MatchEndpoint\Enums\Converters\CapturedPointConverter.cs" />
     <Compile Include="MatchEndpoint\Mastery.cs" />
-    <Compile Include="MatchEndpoint\Converters\CapturedPoint.cs" />
+    <Compile Include="MatchEndpoint\Enums\CapturedPoint.cs" />
     <Compile Include="MatchEndpoint\MatchReference.cs" />
     <Compile Include="MatchEndpoint\Rune.cs" />
-    <Compile Include="MatchEndpoint\Converters\Season.cs" />
-    <Compile Include="MatchEndpoint\Converters\SeasonConverter.cs" />
+    <Compile Include="MatchEndpoint\Enums\Season.cs" />
+    <Compile Include="MatchEndpoint\Enums\Converters\SeasonConverter.cs" />
     <Compile Include="Misc\DateTimeConverterFromString.cs" />
     <Compile Include="Misc\DateTimeConverterFromStringTimestamp.cs" />
     <Compile Include="Misc\GameMode.cs" />
@@ -107,22 +107,22 @@
     <Compile Include="LeagueEndpoint\LeagueEntry.cs" />
     <Compile Include="LeagueEndpoint\MiniSeries.cs" />
     <Compile Include="MatchEndpoint\BannedChampion.cs" />
-    <Compile Include="MatchEndpoint\Converters\BuildingType.cs" />
-    <Compile Include="MatchEndpoint\Converters\BuildingTypeConverter.cs" />
+    <Compile Include="MatchEndpoint\Enums\BuildingType.cs" />
+    <Compile Include="MatchEndpoint\Enums\Converters\BuildingTypeConverter.cs" />
     <Compile Include="MatchEndpoint\Event.cs" />
-    <Compile Include="MatchEndpoint\Converters\EventType.cs" />
-    <Compile Include="MatchEndpoint\Converters\EventTypeConverter.cs" />
+    <Compile Include="MatchEndpoint\Enums\EventType.cs" />
+    <Compile Include="MatchEndpoint\Enums\Converters\EventTypeConverter.cs" />
     <Compile Include="MatchEndpoint\Frame.cs" />
-    <Compile Include="MatchEndpoint\Converters\Lane.cs" />
-    <Compile Include="MatchEndpoint\Converters\LaneConverter.cs" />
-    <Compile Include="MatchEndpoint\Converters\LaneType.cs" />
-    <Compile Include="MatchEndpoint\Converters\LaneTypeConverter.cs" />
-    <Compile Include="MatchEndpoint\Converters\LevelUpType.cs" />
-    <Compile Include="MatchEndpoint\Converters\LevelUpTypeConverter.cs" />
+    <Compile Include="MatchEndpoint\Enums\Lane.cs" />
+    <Compile Include="MatchEndpoint\Enums\Converters\LaneConverter.cs" />
+    <Compile Include="MatchEndpoint\Enums\LaneType.cs" />
+    <Compile Include="MatchEndpoint\Enums\Converters\LaneTypeConverter.cs" />
+    <Compile Include="MatchEndpoint\Enums\LevelUpType.cs" />
+    <Compile Include="MatchEndpoint\Enums\Converters\LevelUpTypeConverter.cs" />
     <Compile Include="MatchEndpoint\MatchDetail.cs" />
     <Compile Include="MatchEndpoint\MatchSummary.cs" />
-    <Compile Include="MatchEndpoint\Converters\MonsterType.cs" />
-    <Compile Include="MatchEndpoint\Converters\MonsterTypeConverter.cs" />
+    <Compile Include="MatchEndpoint\Enums\MonsterType.cs" />
+    <Compile Include="MatchEndpoint\Enums\Converters\MonsterTypeConverter.cs" />
     <Compile Include="MatchEndpoint\Participant.cs" />
     <Compile Include="MatchEndpoint\ParticipantFrame.cs" />
     <Compile Include="MatchEndpoint\ParticipantIdentity.cs" />
@@ -131,16 +131,16 @@
     <Compile Include="MatchEndpoint\ParticipantTimelineData.cs" />
     <Compile Include="MatchEndpoint\Player.cs" />
     <Compile Include="MatchEndpoint\Position.cs" />
-    <Compile Include="MatchEndpoint\Converters\QueueType.cs" />
-    <Compile Include="MatchEndpoint\Converters\QueueTypeConverter.cs" />
-    <Compile Include="MatchEndpoint\Converters\Role.cs" />
-    <Compile Include="MatchEndpoint\Converters\RoleConverter.cs" />
+    <Compile Include="MatchEndpoint\Enums\QueueType.cs" />
+    <Compile Include="MatchEndpoint\Enums\Converters\QueueTypeConverter.cs" />
+    <Compile Include="MatchEndpoint\Enums\Role.cs" />
+    <Compile Include="MatchEndpoint\Enums\Converters\RoleConverter.cs" />
     <Compile Include="MatchEndpoint\Team.cs" />
     <Compile Include="MatchEndpoint\Timeline.cs" />
-    <Compile Include="MatchEndpoint\Converters\TowerType.cs" />
-    <Compile Include="MatchEndpoint\Converters\TowerTypeConverter.cs" />
-    <Compile Include="MatchEndpoint\Converters\WardType.cs" />
-    <Compile Include="MatchEndpoint\Converters\WardTypeConverter.cs" />
+    <Compile Include="MatchEndpoint\Enums\TowerType.cs" />
+    <Compile Include="MatchEndpoint\Enums\Converters\TowerTypeConverter.cs" />
+    <Compile Include="MatchEndpoint\Enums\WardType.cs" />
+    <Compile Include="MatchEndpoint\Enums\Converters\WardTypeConverter.cs" />
     <Compile Include="Misc\Language.cs" />
     <Compile Include="Misc\Platform.cs" />
     <Compile Include="Misc\PlatformConverter.cs" />
