+++ RiotSharpTest/TournamentRiotApiTest.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using RiotSharp;
-using RiotSharp.MatchEndpoint;
+using RiotSharp.MatchEndpoint.Enums;
 using RiotSharp.TournamentEndpoint;
 using System;
 using System.Collections.Generic;
